Reorganize pages and add many pages with new style
